Build on Windows
Based on changes to the quantlib recipe such that it builds a static
library for the dynamic runtime, it is possible to build a python
package on Windows.

However we need to also run swig (by running the setup.py wrap command)
first.

Since quantlib.hpp is now generated correctly by the quantlib recipe it
can be removed from this recipe.